requirements.txt
# Core Flask
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
PyMySQL==1.1.0
cryptography==41.0.4
SQLAlchemy==2.0.21

# Authentication & Security
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
bcrypt==4.0.1

# Session & Cache
Flask-Session==0.5.0
Flask-Caching==2.0.2
redis==5.0.1

# Forms & Validation
Flask-WTF==1.2.1
WTForms==3.0.1
email-validator==2.0.0

# Rate Limiting
Flask-Limiter==3.5.0
limits==3.6.0

# Utilities
python-dotenv==1.0.0
marshmallow==3.20.1
Flask-Cors==4.0.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2

# Production
gunicorn==21.2.0

